Organize Modal provider authentication
